Modification liste

+++ prog_list.py
@@ -58,30 +58,35 @@ def afficher_liste():
 
     return success
 
+def modifier_liste():
+     if os.path.exists("liste"):
+         nom_liste = input("Quelle liste voulez-vous modifier : ").strip() #strip permet de supprimer les espaces inutiles
+         liste_path = os.path.join("liste", f"{nom_liste}.txt")
 
-def modifier_liste(nom_liste):
-    liste_path = os.path.join("liste", f"{nom_liste}.txt")
+         if os.path.exists(liste_path):
+             print(f"Contenu actuel de la liste '{nom_liste}' :\n ")
 
-    if os.path.exists(liste_path):
-        print(f"Contenu actuel de la liste '{nom_liste}' :\n ")
+             # Lire le contenu actuel du fichier
+             with open(liste_path, "r") as fichier_liste:
+                     contenu = fichier_liste.read()
+                     print(contenu)
 
-        # Lire le contenu actuel du fichier
-        with open(liste_path, "r") as fichier_liste:
-                contenu = fichier_liste.read()
-                print(contenu)
+             # Demander à l'utilisateur de saisir la modification
+             nouveau_contenu = input(f"Entrez la modification pour la liste '{nom_liste}': \n")
 
-        # Demander à l'utilisateur de saisir la modification
-        nouveau_contenu = input(f"Entrez la modification pour la liste '{nom_liste}': \n")
+             # Écrire le nouveau contenu dans le fichier
+             with open(liste_path, "w") as fichier_liste:
+                 fichier_liste.write(nouveau_contenu)
 
-        # Écrire le nouveau contenu dans le fichier
-        with open(liste_path, "w") as fichier_liste:
-            fichier_liste.write(nouveau_contenu)
+             print(f"Le contenu de la liste '{nom_liste}' a été modifiée avec succès.")
 
-        print(f"Le contenu de la liste '{nom_liste}' a été modifiée avec succès.")
+         else :
+             print(f"La liste '{nom_liste}' n'existe pas. Vous devez d'abord créer des fichiers.")
 
-    else :
-        print(f"La liste '{nom_liste}' n'existe pas. Vous devez d'abord créer des fichiers.")
+     else :
+         print("Pour modifier, il faut d'abord créer des fichiers.")
 
+Ò
 
 def supprimer_liste():
     if os.path.exists("liste"):
